backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-ai-slim[openrouter]>=0.1.0
chess>=1.11.0
python-dotenv>=1.0.0
rich>=14.0.0
sqlmodel>=0.0.22
aiosqlite>=0.20.0
httpx>=0.28.0
websockets>=14.0
